pyproject.toml
[tool.poetry]
name = "dw-etl-end-to-end"
version = "0.1.0"
description = ""
authors = ["lksprado <lks-prado@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
yfinance = "^0.2.43"
dbt-postgres = "^1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
